Built the groundwork for the app.
App.java holds the data for launching the app.
Dungeon.java generates a map for the player to
roam around in and populates it with the player
and keys.
Key.java generates a key object in the Dungeon
Player,java generates a player object in the
Dungeon that the user has control over.